requirements.txt.txt
streamlit
opencv-python
pytesseract
pillow
pandas
numpy